Allow trivial_numeric_casts in derived code
Simple enum expressions may have no explicit type, like `A = 1`, so then
the derived `1 as i64` is inferred like `1i64 as i64`, a trivial numeric
cast.  But it's quite possible that other expressions could have
explicit types, so we can't just assign it directly either.  The simple
solution is to just allow the `trivial_numeric_casts` in derived code.